packets c.

+++ Packets/S2C/0xAB-ENABLE_FOW.js
@@ -2,6 +2,6 @@ var BasePacket = require('../BasePacket');
 
 module.exports = class extends BasePacket {//S2C.
 	struct = {
-		bitfield_Enabled: 'float',
+		bitfield_Enabled: 'uint8',
 	}
 };
